+++ ohsome_quality_api/indicators/minimal/indicator.py
@@ -23,9 +23,11 @@ class Minimal(BaseIndicator):
         )
 
     def calculate(self) -> None:
-        description = Template(self.metadata.result_description).substitute()
+        description = Template(self.templates.result_description).substitute()
         self.result.value = 1.0
-        self.result.description = description + self.metadata.label_description["green"]
+        self.result.description = (
+            description + self.templates.label_description["green"]
+        )
 
     def create_figure(self) -> None:
         # Do nothing ...