+++ tests/unittests/test_definitions.py
@@ -1,7 +1,5 @@
 import unittest
 
-import pytest
-
 from ohsome_quality_api import definitions
 from ohsome_quality_api.indicators.models import (
     IndicatorMetadata as IndicatorMetadata,
@@ -24,29 +22,3 @@ class TestDefinitions(unittest.TestCase):
         )
 
         self.assertRaises(AssertionError, definitions.get_attribution, ["MSO"])
-
-
-def test_load_metadata_indicator():
-    metadata = definitions.load_metadata("indicators")
-    assert isinstance(metadata, dict)
-    for v in metadata.values():
-        assert isinstance(v, IndicatorMetadata)
-
-
-def test_load_metadata_wrong_module():
-    with pytest.raises(AssertionError):
-        definitions.load_metadata("foo")
-    with pytest.raises(AssertionError):
-        definitions.load_metadata("")
-
-
-def test_get_metadata_indicator():
-    metadata = definitions.get_metadata("indicators", "Minimal")
-    assert isinstance(metadata, IndicatorMetadata)
-
-
-def test_get_metadata_wrong_class():
-    with pytest.raises(KeyError):
-        definitions.get_metadata("indicators", "foo")
-    with pytest.raises(KeyError):
-        definitions.get_metadata("indicators", "")